requirements.txt
PyYAML
rapidfuzz
scrapling
pandas
